requirements.txt
ics>=0.7
tabula-py>=2.1.0
